Reorganise into fs/ directory
